requirements.txt
numpy==1.17.3
phe==1.4.0
torchvision==0.4.1
matplotlib==3.1.1
torch==1.3.0
pycryptodomex==3.9.7
